cli: Make log file and log level configurable via CLI options.
