+++ cli/cmd/common.go
@@ -1,17 +1,60 @@
 package cmd
 
 import (
+	"io"
 	"os"
+	"strings"
 
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
 	"github.com/gluster/glusterd2/pkg/restclient"
 )
 
 var client *restclient.Client
+var logWriter io.WriteCloser
 
 func initRESTClient(hostname string, cacert string, insecure bool) {
 	client = restclient.New(hostname, "", "", cacert, insecure)
+}
+
+func openLogFile(filepath string) (io.WriteCloser, error) {
+	f, err := os.OpenFile(filepath, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	if err != nil {
+		return nil, err
+	}
+	return f, nil
+}
+
+func setLogOutput(w io.Writer) {
+	log.SetOutput(w)
+}
+
+func initLog(logFilePath string, logLevel string) error {
+	// Close the previously opened Log file
+	if logWriter != nil {
+		logWriter.Close()
+		logWriter = nil
+	}
+
+	l, err := log.ParseLevel(strings.ToLower(logLevel))
+	if err != nil {
+		setLogOutput(os.Stderr)
+		log.WithError(err).Debug("Failed to parse log level")
+		return err
+	}
+	log.SetLevel(l)
+	log.SetFormatter(&log.TextFormatter{FullTimestamp: true})
+
+	logFile, err := openLogFile(logFilePath)
+	if err != nil {
+		setLogOutput(os.Stderr)
+		log.WithError(err).Debug("Failed to open log file %s", logFilePath)
+		return err
+	}
+	setLogOutput(logFile)
+	logWriter = logFile
+	return nil
 }
 
 func failure(msg string, err int) {
